rpi-eeprom-config: strip flashrom warning from latest image path
rpi-eeprom-update calls "rpi-eeprom-update -l" to find the latest EEPROM image. When flashrom is not installed, rpi-eeprom-update additionally prints a warning about this:
WARNING: flashrom not found. Setting RPI_EEPROM_USE_FLASHROM to 0

This warning is handled as part of the EEPROM image path, which hence cannot be found.

This is one quick & dirty solution: The output of rpi-eeprom-update is split into a list per line, and only the last line is used for the path.

There are surely better solutions, like using STDOUT only, while printing the warning to STDERR, or skipping the flashrom check for "rpi-eeprom-update -l", as it is not needed to list available images, to assure it always only prints a file path and nothing else.

Signed-off-by: MichaIng <micha@dietpi.com>